bot_telegram_personal_finance/requirements.txt
python-telegram-bot==20.3
pyzbar 
Pillow
# sqlite3
# uuid
# os
# extract nfce
requests
BeautifulSoup4
lxml
dotenv
